Update README.md
For this lab project:

Open a terminal window inside of your Linux virtual machine (VM).
Explore the commands introduced in Module 1.
Inside of your user's home directory, create a directory structure with files that replicates what is found in the following screenshot.
You will need to create the homework and labs directories and the drafts subdirectory inside of the homework directory with the two empty files inside of each directory and subdirectory.
Your shell prompt should be different than the one in the screenshot shared below (a different username and hostname).
Your output for the pwd command should be different than the one in the screenshot shared below.
IFT 250 Lab 1 Screenshot example

Submit in a screenshot (.jpg or .png file format) of a pwd and a ls -alR homework labs command output that was executed from within your user's home directory on your system (similar to the above example).
Your screenshot(s) must clearly show a shell prompt that has a username comprised of your first name (or first initial) and last name. The username in the shell prompt must match your first name and last name displayed in your Canvas profile.